+++ structc/struct/tstr.h
@@ -6,8 +6,8 @@
 #ifndef _STRUCT_TSTR
 
 struct tstr {
-    size_t len;     // 长度
-    size_t cap;     // 容量
+    unsigned len;   // 长度
+    unsigned cap;   // 容量
     char * str;     // 字符池
 };
 
@@ -32,7 +32,7 @@ free((var)->str)
 // len		: 创建串的长度
 // return	: 返回创建好的字符串,内存不足会打印日志退出程序
 //
-extern tstr_t tstr_create(const char * str, size_t len);
+extern tstr_t tstr_create(const char * str, unsigned len);
 extern tstr_t tstr_creates(const char * str);
 
 //
@@ -50,7 +50,7 @@ extern void tstr_delete(tstr_t tstr);
 //
 extern void tstr_appendc(tstr_t tstr, int c);
 extern void tstr_appends(tstr_t tstr, const char * str);
-extern void tstr_appendn(tstr_t tstr, const char * str, size_t sz);
+extern void tstr_appendn(tstr_t tstr, const char * str, unsigned sz);
 
 //
 // tstr_cstr - 通过cstr_t串得到一个c的串以'\0'结尾
@@ -65,7 +65,7 @@ extern char * tstr_cstr(tstr_t tstr);
 // len		: 扩容的长度
 // return	: tstr->str + tstr->len 位置的串
 //
-char * tstr_expand(tstr_t tstr, size_t len);
+char * tstr_expand(tstr_t tstr, unsigned len);
 
 //
 // tstr_popup - 从字符串头弹出len长度字符
@@ -73,7 +73,7 @@ char * tstr_expand(tstr_t tstr, size_t len);
 // len		: 弹出的长度
 // return	: void
 //
-extern void tstr_popup(tstr_t tstr, size_t len);
+extern void tstr_popup(tstr_t tstr, unsigned len);
 
 //
 // tstr_dupstr - 得到一个精简的c的串, 需要自己事后free
